+++ src/main/java/org/ahpuh/surf/category/domain/Category.java
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import org.ahpuh.surf.common.entity.BaseEntity;
 import org.ahpuh.surf.post.entity.Post;
-import org.ahpuh.surf.user.entity.User;
+import org.ahpuh.surf.user.domain.User;
 import org.hibernate.annotations.Formula;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.Where;
