+++ src/main/java/org/ahpuh/surf/category/domain/CategoryRepository.java
@@ -1,7 +1,6 @@
 package org.ahpuh.surf.category.domain;
 
-import org.ahpuh.surf.category.entity.Category;
-import org.ahpuh.surf.user.entity.User;
+import org.ahpuh.surf.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
