+++ src/main/java/org/ahpuh/surf/like/converter/LikeConverter.java
@@ -2,7 +2,7 @@ package org.ahpuh.surf.like.converter;
 
 import org.ahpuh.surf.like.entity.Like;
 import org.ahpuh.surf.post.entity.Post;
-import org.ahpuh.surf.user.entity.User;
+import org.ahpuh.surf.user.domain.User;
 import org.springframework.stereotype.Component;
 
 @Component
