+++ src/main/java/org/ahpuh/surf/like/entity/Like.java
@@ -2,7 +2,7 @@ package org.ahpuh.surf.like.entity;
 
 import lombok.*;
 import org.ahpuh.surf.post.entity.Post;
-import org.ahpuh.surf.user.entity.User;
+import org.ahpuh.surf.user.domain.User;
 
 import javax.persistence.*;
