+++ src/main/java/org/ahpuh/surf/post/repository/PostRepository.java
@@ -2,7 +2,7 @@ package org.ahpuh.surf.post.repository;
 
 import org.ahpuh.surf.category.domain.Category;
 import org.ahpuh.surf.post.entity.Post;
-import org.ahpuh.surf.user.entity.User;
+import org.ahpuh.surf.user.domain.User;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
