+++ src/main/java/org/ahpuh/surf/post/repository/PostRepositoryQuerydsl.java
@@ -3,7 +3,7 @@ package org.ahpuh.surf.post.repository;
 import org.ahpuh.surf.post.dto.ExploreDto;
 import org.ahpuh.surf.post.dto.PostCountDto;
 import org.ahpuh.surf.post.dto.PostScoreCategoryDto;
-import org.ahpuh.surf.user.entity.User;
+import org.ahpuh.surf.user.domain.User;
 import org.springframework.data.domain.Pageable;
 
 import java.time.LocalDate;
