+++ src/main/java/org/ahpuh/surf/user/service/UserService.java
@@ -6,15 +6,15 @@ import org.ahpuh.surf.follow.repository.FollowRepository;
 import org.ahpuh.surf.jwt.JwtAuthentication;
 import org.ahpuh.surf.jwt.JwtAuthenticationToken;
 import org.ahpuh.surf.s3.S3Service;
-import org.ahpuh.surf.user.converter.UserConverter;
+import org.ahpuh.surf.user.domain.User;
+import org.ahpuh.surf.user.domain.UserConverter;
+import org.ahpuh.surf.user.domain.UserRepository;
 import org.ahpuh.surf.user.dto.request.UserJoinRequestDto;
 import org.ahpuh.surf.user.dto.request.UserUpdateRequestDto;
 import org.ahpuh.surf.user.dto.response.UserFindInfoResponseDto;
 import org.ahpuh.surf.user.dto.response.UserJoinResponseDto;
 import org.ahpuh.surf.user.dto.response.UserLoginResponseDto;
 import org.ahpuh.surf.user.dto.response.UserUpdateResponseDto;
-import org.ahpuh.surf.user.entity.User;
-import org.ahpuh.surf.user.repository.UserRepository;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.password.PasswordEncoder;
