+++ src/test/java/org/ahpuh/surf/integration/post/PostTest.java
@@ -8,9 +8,9 @@ import org.ahpuh.surf.post.entity.Post;
 import org.ahpuh.surf.post.repository.PostRepository;
 import org.ahpuh.surf.post.service.PostService;
 import org.ahpuh.surf.user.controller.UserController;
+import org.ahpuh.surf.user.domain.User;
+import org.ahpuh.surf.user.domain.UserRepository;
 import org.ahpuh.surf.user.dto.request.UserJoinRequestDto;
-import org.ahpuh.surf.user.entity.User;
-import org.ahpuh.surf.user.repository.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
